+++ src/vmm/src/builder.rs
@@ -905,6 +905,8 @@ pub(crate) mod tests {
                 ),
                 rate_limiter: None,
                 file_engine_type: None,
+                blk_size: None,
+                topology: None,
 
                 socket: None,
             };