+++ src/vmm/src/device_manager/persist.rs
@@ -828,6 +828,13 @@ mod tests {
       "path_on_host": "{}",
       "rate_limiter": null,
       "io_engine": "Sync",
+      "blk_size": 512,
+      "topology": {{
+        "physical_block_exp": 0,
+        "alignment_offset": 0,
+        "min_io_size": 0,
+        "opt_io_size": 128
+      }},
       "socket": null
     }}
   ],