+++ CHANGELOG.md
@@ -34,6 +34,10 @@ and this project adheres to
   the `huge_pages` field to `PUT /snapshot/load`, allowing the restored microVM
   to reuse the snapshot's host page configuration or select `None`,
   `Transparent`, or `2M`.
+- [#6098](https://github.com/firecracker-microvm/firecracker/pull/6098): Added
+  new VIRTIO_BLK_F_BLK_SIZE and VIRTIO_BLK_F_TOPOLOGY features to the
+  virtio-block device. More information is in the new [block](docs/block.md)
+  documentation.
 
 ### Changed